Using `chalk` in runtime (#1640)
Some terminals may not support colors and I'd like to avoid assumptions
on that issue.
I tried several libraries, but in the end Chalk is the most reliable
one, because it has fallbacks.

Using latest Chalk v5 (which is ESM only) requires to make
`createLogger()` fn async.
