+++ tests/unit/startup-logo.spec.ts
@@ -1,10 +1,11 @@
 import { getStartupLogo } from "../../src/startup-logo";
 import { describe, expect, test } from "vitest";
+import chalk from "chalk";
 
 describe("Startup logo", () => {
   describe("getStartupLogo()", () => {
     test("should return the logo", () => {
-      expect(getStartupLogo().split("\n").length).toBe(14);
+      expect(getStartupLogo(chalk).split("\n").length).toBe(14);
     });
   });
 });
